split test_typedb_interface into 2 test files
