+++ ros_typedb/test/test_typedb_interface_unit.py
@@ -573,50 +573,6 @@ def test_init_loads_schema_for_new_database_when_reload_disabled(monkeypatch):
     assert loaded_schema_paths == ['schema.tql']
 
 
-@pytest.fixture
-def typedb_interface():
-    typedb_interface = TypeDBInterface(
-        'localhost:1729',
-        'test_database',
-        force_database=True,
-        schema_path=['test/typedb_test_data/schema.tql'],
-        data_path=['test/typedb_test_data/data.tql'],
-        force_data=True,
-        sort_fetch_results=True
-    )
-    yield typedb_interface
-    typedb_interface.delete_database()
-
-
-def _fetch_entities(typedb_interface):
-    return typedb_interface.get_aggregate_database(
-        'match $e isa entity; get $e; count;')
-
-
-def _fetch_relations(typedb_interface):
-    return typedb_interface.get_aggregate_database(
-        'match $r isa relation; get $r; count;')
-
-
-def _fetch_attributes(typedb_interface):
-    return typedb_interface.get_aggregate_database(
-        'match $a isa attribute; get $a; count;')
-
-
-def test_delete_all_data_removes_entities_relations_and_attributes(
-        typedb_interface):
-    """delete_all_data removes all live fixture data."""
-    assert _fetch_entities(typedb_interface)
-    assert _fetch_relations(typedb_interface)
-    assert _fetch_attributes(typedb_interface)
-
-    typedb_interface.delete_all_data()
-
-    assert _fetch_entities(typedb_interface) == 0
-    assert _fetch_relations(typedb_interface) == 0
-    assert _fetch_attributes(typedb_interface) == 0
-
-
 def test_delete_all_data_raises_when_delete_step_fails(monkeypatch):
     """delete_all_data stops instead of loading data over residual data."""
     typedb_interface = TypeDBInterface.__new__(TypeDBInterface)
@@ -634,30 +590,6 @@ def test_delete_all_data_raises_when_delete_step_fails(monkeypatch):
         typedb_interface.delete_all_data()
 
     assert queries == ['match $e isa entity; delete $e isa entity;']
-
-
-def test_delete_entities_before_relations_leaves_no_residual_data(
-        typedb_interface):
-    """Document TypeDB behavior for the current delete_all_data order."""
-    assert _fetch_relations(typedb_interface)
-
-    typedb_interface.delete_from_database(
-        'match $e isa entity; delete $e isa entity;')
-    after_entity_delete_error = typedb_interface.last_error
-    entities_after_entity_delete = _fetch_entities(typedb_interface)
-    relations_after_entity_delete = _fetch_relations(typedb_interface)
-
-    typedb_interface.delete_from_database(
-        'match $r isa relation; delete $r isa relation;')
-    typedb_interface.delete_from_database(
-        'match $a isa attribute; delete $a isa attribute;')
-
-    assert after_entity_delete_error == ''
-    assert entities_after_entity_delete == 0
-    assert relations_after_entity_delete == 0
-    assert _fetch_entities(typedb_interface) == 0
-    assert _fetch_relations(typedb_interface) == 0
-    assert _fetch_attributes(typedb_interface) == 0
 
 
 def test_convert_py_type_to_query_type_escapes_strings():
@@ -681,12 +613,6 @@ def test_delete_thing_uses_query_type_conversion(monkeypatch):
     assert typedb_interface.delete_thing('person', 'name', "O'Brien") is True
     assert "has name 'O\\'Brien'" in captured_query
     assert 'has name "O\'Brien"' not in captured_query
-
-
-def test_load_bad_data_raises(typedb_interface):
-    """Regression: load_data must raise on invalid TypeQL, not silently pass."""
-    with pytest.raises(Exception):
-        typedb_interface.load_data('test/typedb_test_data/bad_data.tql')
 
 
 def test_driver_connection_timeout(monkeypatch):
@@ -743,62 +669,6 @@ def test_driver_connection_timeout_closes_late_driver(monkeypatch):
     assert late_driver.closed
 
 
-def test_create_and_delete_database():
-    typedb_interface = TypeDBInterface(
-        'localhost:1729',
-        'test_database',
-        force_database=True,
-        schema_path=['test/typedb_test_data/schema.tql'],
-        data_path=['test/typedb_test_data/data.tql'],
-        force_data=True,
-    )
-
-    assert typedb_interface.driver.databases.contains('test_database')
-
-    typedb_interface.delete_database()
-    assert not typedb_interface.driver.databases.contains('test_database')
-
-
-def test_define_query(typedb_interface):
-    assert typedb_interface.define_database('define MyEntity sub entity;')
-    assert typedb_interface.define_database(
-        'define MyEntity aa entity') is None
-
-
-def test_insert_entity(typedb_interface):
-    result_insert = typedb_interface.insert_entity(
-        'person', [('email', 'test@email.test'), ('nickname', 't')])
-    assert result_insert is not None
-    query = """
-        match $entity isa person,
-        has email "test@email.test",
-        has nickname "t";
-        get $entity;
-        count;
-    """
-    result = typedb_interface.get_aggregate_database(query)
-    assert result > 0
-
-    result_insert = typedb_interface.insert_entity(
-        'something_wrong', [('email', 'test@email.test'), ('nickname', 't')])
-    assert result_insert is None
-
-
-def test_delete_thing(typedb_interface):
-    typedb_interface.insert_entity('person', [('email', 'test@email.test')])
-    typedb_interface.delete_thing('person', 'email', 'test@email.test')
-    query = """
-        match $entity isa person, has email "test@email.test";
-        get $entity;
-    """
-    result = typedb_interface.get_database(query)
-    assert len(result) == 0
-
-    wrong_result = typedb_interface.delete_thing(
-        'something_wrong', 'email', 'test@email.test')
-    assert wrong_result is None
-
-
 @pytest.mark.parametrize('key_value, expected_query_value', [
     ('test@email.test', "'test@email.test'"),
     (33, '33'),
@@ -819,420 +689,6 @@ def test_delete_thing_formats_key_value_for_type(key_value, expected_query_value
 
     assert typedb_interface.delete_thing('person', 'email', key_value) is True
     assert f'has email {expected_query_value};' in queries[0]
-
-
-@pytest.mark.parametrize('attr, attr_value', [
-    ('nickname', 't'),
-    ('alive', True),
-    ('age', 33),
-    ('height', 3.237),
-    ('birth-date', datetime.now()),
-])
-def test_insert_attribute_in_thing(typedb_interface, attr, attr_value):
-    typedb_interface.insert_entity('person', [('email', 'test@email.test')])
-    typedb_interface.insert_attribute_in_thing(
-        'person', 'email', 'test@email.test', attr, attr_value)
-    result = typedb_interface.fetch_attribute_from_thing(
-        'person', [('email', 'test@email.test')], attr)
-    assert result
-
-
-@pytest.mark.parametrize('attr, attr_value', [
-    ('nickname', 't'),
-    ('alive', True),
-    ('age', 33),
-    ('height', 3.237),
-    ('birth-date', datetime.fromisoformat(
-        datetime.now().isoformat(timespec='milliseconds'))),
-])
-def test_fetch_attribute_from_thing(typedb_interface, attr, attr_value):
-    typedb_interface.insert_entity(
-        'person',
-        [('email', 'test@email.test'), ('gender', 'male')]
-    )
-    typedb_interface.insert_attribute_in_thing(
-        'person', 'email', 'test@email.test', attr, attr_value)
-
-    result = typedb_interface.fetch_attribute_from_thing(
-        'person', [('email', 'test@email.test'), ('gender', 'male')], attr)
-
-    assert result[0] == attr_value
-
-
-@pytest.mark.parametrize('attr, attr_value', [
-    ('nickname', 't'),
-    ('alive', True),
-    ('age', 33),
-    ('height', 3.237),
-    ('birth-date', datetime.fromisoformat(
-        datetime.now().isoformat(timespec='milliseconds'))),
-])
-def test_delete_attribute_from_thing(typedb_interface, attr, attr_value):
-    typedb_interface.insert_entity('person', [('email', 'test@email.test')])
-    typedb_interface.insert_attribute_in_thing(
-        'person', 'email', 'test@email.test', attr, attr_value)
-
-    typedb_interface.delete_attribute_from_thing(
-        'person', 'email', 'test@email.test', attr)
-
-    result = typedb_interface.fetch_attribute_from_thing(
-        'person', [('email', 'test@email.test')], attr)
-
-    assert len(result) == 0
-
-
-@pytest.mark.parametrize('attr, attr_value, new_v', [
-    ('nickname', 't', 'new_t'),
-    ('alive', True, False),
-    ('age', 33, 56),
-    ('height', 3.237, 1.66),
-    ('birth-date', datetime.fromisoformat(
-        datetime.now().isoformat(timespec='milliseconds')),
-        datetime.fromisoformat(
-            datetime.now().isoformat(timespec='milliseconds'))),
-])
-def test_update_attribute_in_thing(typedb_interface, attr, attr_value, new_v):
-    typedb_interface.insert_entity('person', [('email', 'test@email.test')])
-    typedb_interface.insert_attribute_in_thing(
-        'person', 'email', 'test@email.test', attr, attr_value)
-    result_update = typedb_interface.update_attribute_in_thing(
-        'person', 'email', 'test@email.test', attr, new_v)
-    result = typedb_interface.fetch_attribute_from_thing(
-        'person', [('email', 'test@email.test')], attr)
-
-    assert result_update is not None and result[0] == new_v
-
-
-def test_insert_relationship(typedb_interface):
-    typedb_interface.insert_entity('person', [('email', 'test@email.test')])
-    typedb_interface.insert_entity('person', [('email', 'test2@email.test')])
-    typedb_interface.insert_entity('person', [('email', 'test3@email.test')])
-    typedb_interface.insert_entity('person', [('email', 'test4@email.test')])
-
-    related_things_dict = {
-        'employee': [
-            ('person', 'email', 'test@email.test'),
-            ('person', 'email', 'test2@email.test')],
-        'employer': [
-            ('person', 'email', 'test3@email.test'),
-            ('person', 'email', 'test4@email.test')
-        ]
-    }
-    attribute_list = [
-        ('salary', 2333), ('role-name', 'boss'), ('role-name', 'super boss')]
-    typedb_interface.insert_relationship(
-        'employment', related_things_dict, attribute_list)
-
-    query = """
-        match $r (employee:$ee, employer:$er)isa employment,
-        has salary 2333,
-        has role-name "boss",
-        has role-name "super boss";
-        get $r;
-        count;
-    """
-    result = typedb_interface.get_aggregate_database(query)
-    assert result > 0
-
-
-@pytest.mark.parametrize('things_dict', [
-    ({
-        'person': [
-            {
-                'prefix': 'p1',
-                'attributes': {
-                    'email': 'test@email.test',
-                    'nickname': 't',
-                    'height': 1.80,
-                    'age': 18,
-                    'alive': True,
-                    'birth-date': datetime.now()
-                }
-            },
-            {
-                'prefix': 'p2',
-                'attributes': {
-                    'email': 'test2@email.test',
-                    'nickname': 't2',
-                    'height': 1.33,
-                    'age': 128,
-                    'alive': False,
-                    'birth-date': datetime.now()
-                }
-            },
-        ],
-        'robot': [
-            {
-                'prefix': 'r',
-                'attributes': {
-                    'full-name': 'robot123',
-                    'height': 1.0,
-                    'age': 0,
-                    'alive': True,
-                    'birth-date': datetime.now()
-                }
-            },
-        ],
-        'employment': [
-            {
-                'prefix': 'e',
-                'attributes': {
-                    'salary': 2333,
-                    'role-name': ['boss', 'super boss'],
-                },
-                'relationship': {
-                    'employee': 'p1',
-                    'employer': 'p2'
-                }
-            },
-        ]
-    }),
-])
-def test_dict_to_query(typedb_interface, things_dict):
-    query = typedb_interface.dict_to_query(things_dict)
-    insert_result = typedb_interface.insert_database('insert ' + query)
-    assert insert_result is not None
-
-
-@pytest.mark.parametrize('match_dict, r_dict', [
-    (
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test_person@test.test',
-                    },
-                    'insert_attributes': {
-                        'nickname': 't',
-                        'height': 1.80,
-                        'age': 18,
-                        'alive': True,
-                        'birth-date': datetime.now()
-                    }
-                },
-            ],
-        },
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test_person@test.test',
-                        'nickname': 't',
-                        'height': 1.80,
-                        'age': 18,
-                        'alive': True,
-                        'birth-date': datetime.now()
-                    }
-                },
-            ],
-        },
-    ),
-])
-def test_insert_attributes(typedb_interface, match_dict, r_dict):
-    r = typedb_interface.insert_attributes_in_thing(match_dict)
-    query = typedb_interface.dict_to_query(r_dict)
-    match_result = typedb_interface.get_aggregate_database(
-        'match ' + query + 'get; count;')
-    assert r is not None and r is not False and match_result > 0
-
-
-@pytest.mark.parametrize('insert_dict, match_dict', [
-    (
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test@test.test',
-                        'nickname': 't',
-                        'height': 1.80,
-                        'age': 18,
-                        'alive': True,
-                        'birth-date': datetime.now()
-                    }
-                },
-            ],
-            'robot': [
-                {
-                    'prefix': 'r',
-                    'attributes': {
-                        'full-name': 'robot123',
-                        'height': 1.0,
-                        'age': 0,
-                        'alive': True,
-                        'birth-date': datetime.now(),
-                        'robot-type': 'auv'
-                    }
-                },
-            ],
-        },
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test@test.test',
-                    },
-                    'delete_attributes': [
-                        'nickname', 'height', 'age', 'alive', 'birth-date']
-                },
-            ],
-            'robot': [
-                {
-                    'attributes': {
-                        'full-name': 'robot123',
-                    },
-                    'delete_attributes': ['robot-type']
-                },
-            ],
-        },
-    )
-])
-def test_delete_attributes(
-        typedb_interface, insert_dict, match_dict):
-
-    query = typedb_interface.dict_to_query(insert_dict)
-    insert_result = typedb_interface.insert_database('insert ' + query)
-    assert insert_result is not None
-    assert len(insert_result) > 0
-
-    delete_result = typedb_interface.delete_attributes_from_thing(match_dict)
-    assert delete_result is True
-
-
-@pytest.mark.parametrize('insert_dict, update_dict, r_dict', [
-    (
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test@test.test',
-                        'nickname': 't',
-                        'height': 1.80,
-                        'age': 18,
-                        'alive': True,
-                        'birth-date': datetime.now()
-                    }
-                },
-            ],
-        },
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test@test.test',
-                    },
-                    'update_attributes': {
-                        'nickname': 't2',
-                        'height': 1.50,
-                        'age': 17,
-                        'alive': False,
-                    }
-                },
-            ],
-        },
-        {
-            'person': [
-                {
-                    'prefix': 'p1',
-                    'attributes': {
-                        'email': 'test@test.test',
-                        'nickname': 't2',
-                        'height': 1.50,
-                        'age': 17,
-                        'alive': False,
-                    }
-                },
-            ],
-        },
-    ),
-])
-def test_update_attributes(typedb_interface, insert_dict, update_dict, r_dict):
-    query = typedb_interface.dict_to_query(insert_dict)
-    typedb_interface.insert_database('insert ' + query)
-
-    r = typedb_interface.update_attributes_in_thing(update_dict)
-    query = typedb_interface.dict_to_query(r_dict)
-    match_result = typedb_interface.get_aggregate_database(
-        'match ' + query + 'get; count;')
-    assert r is not None and r is not False and match_result > 0
-
-
-def test_get_query(typedb_interface):
-    query = """
-        match
-            $p isa person, has full-name $name, has email $email;
-        get $name, $email;
-        sort $name asc; limit 3;
-    """
-    result = typedb_interface.get_database(query)
-    name = result[0].get('name').as_attribute().get_value()
-    assert len(result) == 3 and name == 'Ahmed Frazier'
-
-
-def test_fetch_query(typedb_interface):
-    query = """
-    match
-        $company isa company, has name "TU Delft";
-    fetch
-        employee_names: {
-            match
-                $e (employer: $company, employee: $employee) isa employment;
-            fetch
-                $employee: full-name, email;
-                $e: salary;
-        };
-    """
-    result = typedb_interface.fetch_database(query)
-    assert len(result) == 1
-    assert len(result[0]['employee_names']) == 3
-
-    expected = [
-        {
-            'salary': 30000,
-            'email': 'phd@tudelft.nl',
-            'full-name': 'PhD candidate 1'
-        },
-        {
-            'salary': 999999,
-            'email': 'boss@tudelft.nl',
-            'full-name': 'Big Boss'
-        },
-        {
-            'salary': 0,
-            'email': 'guest@tudelft.nl',
-            'full-name': 'Random guest'
-        }
-    ]
-
-    actual = []
-    for emp in result[0]['employee_names']:
-        actual.append({
-            'salary': emp['e']['salary'][0]['value'],
-            'email': emp['employee']['email'][0]['value'],
-            'full-name': emp['employee']['full-name'][0]['value']
-        })
-
-    # Compare as sets so order does not matter
-    assert {tuple(sorted(d.items())) for d in actual} == \
-        {tuple(sorted(d.items())) for d in expected}
-
-
-def test_register_method(typedb_interface):
-    def get_name_email(self, name):
-        query = f"""
-            match
-                $p isa person, has full-name '{name}';
-            fetch
-                $p: email;
-        """
-        return typedb_interface.fetch_database(query)[0]['p']['email'][0]['value']
-    typedb_interface.register_method('get_name_email', get_name_email)
-    assert typedb_interface.get_name_email('Big Boss') == 'boss@tudelft.nl'
 
 
 def test_database_query_reconnects_after_failed_health_check(monkeypatch):
